+++ templates/Beef.Template.Solution/content/Company.AppName.Business/Data/EfDb.cs
@@ -1,0 +1,12 @@
+﻿using Beef.Data.EntityFrameworkCore;
+
+namespace Company.AppName.Business.Data
+{
+    /// <summary>
+    /// Represents the <b>Beef.Demo</b> database using Entity Framework.
+    /// </summary>
+    public class EfDb : EfDb<EfDbContext, EfDb>
+    {
+        public EfDb() => OnUpdatePreReadForNotFound = true;
+    }
+}